3ds port: Finished most things but graphics
* The emulator basically works, sorta!

* Rewrote chip8::loadApplication to work on the 3ds

* Ported the main loop: modes, displaying registers

* Very badly implemented the chip 8 display to quickly test if everything was working

* Cycles per second counter and limiting speed of emulation are ported but don't work

* Further improvements to overall system stability and other minor adjustments have been made to enhance the user experience.

+++ octochip-8-3ds/source/chip8.h
@@ -20,7 +20,7 @@ private:
 	unsigned char sound_timer; //Sounds buzzer when 0 is reached
 	unsigned short stack[16]; //Stack
 	unsigned short sp; //Stack pointer
-	unsigned long rom_size; //ROM size
+	unsigned long long rom_size; //ROM size
 
 	void init(); //Initialize data
 };
